Changed from double to float. Added viz part. Fixed bugs.

+++ utils/include/retriever.h
@@ -7,16 +7,16 @@
 typedef struct retriever_t {
   int period;
   int current_step;
-  double *data;
+  float *data;
   decomp_t *dp;
 } retriever_t;
   
 retriever_t *retriever_init(decomp_t *dp, int period);
-void retriever_feed(retriever_t *rp, double *I);
-double *retriever_get_step(retriever_t *rp, int step);
-double *retriever_get_laststep(retriever_t *rp);
-int retriever_get_chunk(retriever_t *rp, int pos, double *data);
+void retriever_feed(retriever_t *rp, float *I);
+float *retriever_get_step(retriever_t *rp, int step);
+float *retriever_get_laststep(retriever_t *rp);
+int retriever_get_chunk(retriever_t *rp, int pos, float *data);
 void retriever_finalize(retriever_t *rp);
-//void retriever_get(retriever_t *rp, int *pos, int pos_count, double *data);
+//void retriever_get(retriever_t *rp, int *pos, int pos_count, float *data);
 
 #endif
