+++ yandex/include/quantizer.h
@@ -5,11 +5,11 @@
 
 void quantizer_init(int level);
 
-void quantizer_restart(double *ct, int size);
+void quantizer_restart(float *ct, int size);
 
-int quantizer_quantize_singleton(double ct);
+int quantizer_quantize_singleton(float ct);
 
-void quantizer_quantize(double *ct, int *data_dc, int size);
+void quantizer_quantize(float *ct, int *data_dc, int size);
 
 void quantizer_finalize();
